app/scraper/scraping_job.py
from typing import Any, Dict, List, Optional
from pyppeteer.page import Page  # Assuming you're using pyppeteer

class ScrapingJob:
    def __init__(self, job: Dict[str, Any], parent_config: Optional[Dict[str, Any]] = None) -> None:
        """
        Initialize a new web-scraper instance.
        """
        self.job = job
        self.config = job.get("config", parent_config)  # Use parent config if no config provided
        self.browser = BrowserInstance(**self.config)  # Please define BrowserInstance

    def scrape_element(self, xpath: str, page: Page) -> Any:
        """
        Scrape the data from a webpage element based on the provided xpath.
        """
        # Implement logic to scrape an element based on the provided xpath

    def click_element(self, xpath: str, page: Page) -> None:
        """
        Click on a webpage element based on the provided xpath.
        """
        # Implement logic to click an element based on the provided xpath

    async def scrape_data(self, scrape_details: List[Dict[str, str]], page:F Page) -> None:
        """
        Scrape data for each item in the scrape_details list.

        :param scrape_details: List of details specifying what to scrape
        :param page: The Pyppeteer page object
        """
        for scrape_info in scrape_details:
            await self.scrape_element(scrape_info["xpath"], page)  # Adjusted parameters

    def perform_actions(self, actions: List[Dict[str, Any]], page: Page) -> None:
        """
        Iterate through actions and execute the corresponding function.

        :param actions: List of actions to perform
        :param page: The Pyppeteer page object
        """
        for action in actions:
            if "job" in action:
                job_type = action["job"].get("type")
                if job_type == "paginator":
                    job = PaginatorJob(**action["job"])  # Pass other necessary arguments
                elif job_type == "extractor":
                    job = ExtractorJob(**action["job"])  # Pass other necessary arguments
                job.perform_actions(page)  # Ensure perform_actions is implemented in job classes
            elif "scrape" in action:
                await self.scrape_data(action["scrape"], page)
            elif "click" in action:
                self.click_element(action["click"]["xpath"], page)  # Adjust to fit the correct structure of your "click" action


    async def execute_job(self):
        """
        Create a browser, navigate to a URL, perform actions, scrape data, and close the browser.
        """
        page = await self.browser.newPage()  # Assuming BrowserInstance class has a newPage method to create a new browser page
        await page.goto(self.job.get('url'))  # Navigate to the URL specified in the job

        actions = self.job.get('actions', [])
        await self.perform_actions(actions, page)  # Perform the actions specified in the job

        # Once actions are performed, close the browser
        await self.browser.close()  # Assuming BrowserInstance class has a close method to close the browser


class PaginatorJob:
    def __init__(self, **kwargs) -> None:
        self.config = kwargs
        self.validate_actions()

    def validate_actions(self) -> None:
        """
        Validate that the actions contain a "paginate" action.
        """
        actions = self.config.get('actions', [])
        if not any(action.get('type') == 'paginate' for action in actions):
            raise ValueError('PaginatorJob must contain a "paginate" action')


    async def perform_actions(self, actions: List[Dict[str, Any]], page: Page) -> None:
        # Perform actions on the first page before entering the loop
        await super().perform_actions(actions, page)

        # Extract the xpath of the 'next' button
        next_button_xpath = actions.get('paginate', {}).get('next_page_button')

        # Loop to paginate through all the pages as long as the 'next' button is found on the page
        while next_button_xpath and await page.querySelector(next_button_xpath):
            # Click the 'next' button to go to the next page
            await page.click(next_button_xpath)

            # Perform the actions on the new page
            await super().perform_actions(actions, page)

class ExtractorJob:
    def __init__(self, **kwargs) -> None:
        self.config = kwargs

    async def perform_actions(self, page: Page) -> None:
        """
        Perform actions specific to an extraction job on the given page.

        :param page: The Pyppeteer page object
        """
        # Implement logic specific to extraction jobs





